alterações

+++ sistema/src/Controller/OrcamentosController.php
@@ -2,6 +2,8 @@
 namespace App\Controller;
 
 use App\Controller\AppController;
+use Cake\Http\Exception\NotFoundException;
+use Cake\Http\Exception\BadRequestException;
 
 /**
  * Orcamentos Controller
@@ -12,35 +14,32 @@ use App\Controller\AppController;
  */
 class OrcamentosController extends AppController
 {
-    /**
-     * Index method
-     *
-     * @return \Cake\Http\Response|null
-     */
-    public function index()
+
+    public function retornaOrcamento($id=null)
     {
-        $this->paginate = [
-            'contain' => ['Projetos']
-        ];
-        $orcamentos = $this->paginate($this->Orcamentos);
+        $this->request->accepts('get');
+        $dados = $_GET;
 
-        $this->set(compact('orcamentos'));
-    }
+        $hash = $this->request->getParam('_csrfToken');
 
-    /**
-     * View method
-     *
-     * @param string|null $id Orcamento id.
-     * @return \Cake\Http\Response|null
-     * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
-     */
-    public function view($id = null)
-    {
-        $orcamento = $this->Orcamentos->get($id, [
-            'contain' => ['Projetos', 'Contratos']
-        ]);
+        if(!isset($dados['hash']) || $dados['hash'] != $hash){
+            throw new BadRequestException();
+        }
 
-        $this->set('orcamento', $orcamento);
+        if(!$id){
+            $retorno = '';
+        }else {
+
+            $orcamento = $this->Orcamentos->get($id, [
+                'contain' => ['Projetos']
+            ]);
+            $orcamento->custo = $orcamento->custo();
+            $orcamento->total = $orcamento->total();
+            $orcamento->projeto->custo_estimado = $orcamento->projeto->custoEstimado();
+            $retorno = json_encode($orcamento);
+        }
+
+        $this->set(compact('retorno'));
     }
 
     /**
@@ -48,45 +47,45 @@ class OrcamentosController extends AppController
      *
      * @return \Cake\Http\Response|null Redirects on successful add, renders view otherwise.
      */
-    public function add()
+    public function add($projeto_id = null)
     {
+        if(empty($projeto_id)){
+            throw new NotFoundException('Selecione um cliente');
+        }
+
+        $user = $sessao = $this->Auth->user();
+        $orcamentos = null;
+
         $orcamento = $this->Orcamentos->newEntity();
-        if ($this->request->is('post')) {
-            $orcamento = $this->Orcamentos->patchEntity($orcamento, $this->request->getData());
-            if ($this->Orcamentos->save($orcamento)) {
-                $this->Flash->success(__('The orcamento has been saved.'));
+        if($projeto_id) {
+            if ($this->request->is('post')) {
+                $dados = $this->request->getData();
 
-                return $this->redirect(['action' => 'index']);
+                $dt = explode('/',$dados['data_inicial']);
+                $dados['data_inicial'] = ($dados['data_inicial']<>''? date('Y-m-d',strtotime($dt[2].'-'.$dt[1].'-'.$dt[0])):null);
+                $dt = explode('/',$dados['data_entrega']);
+                $dados['data_entrega'] = ($dados['data_entrega']<>''? date('Y-m-d',strtotime($dt[2].'-'.$dt[1].'-'.$dt[0])):null);
+                $dados['custo'] = (!empty($dados['custo'])?str_replace(',','.',preg_replace("/[^0-9,]/", "", $dados['custo'])): null);
+                $dados['total'] = (!empty($dados['total'])?str_replace(',','.',preg_replace("/[^0-9,]/", "", $dados['total'])): null);
+                $dados['empresa_id'] = $user['empresa_id'];
+                $dados['u_id'] = $user['id'];
+
+                if (!empty($dados['id'])) {
+                    $orcamento = $this->Orcamentos->get($dados['id']);
+                }
+
+                $orcamento = $this->Orcamentos->patchEntity($orcamento, $dados);
+
+                if ($this->Orcamentos->save($orcamento)) {
+                    //$this->Flash->success(__('The orcamento has been saved.'));
+                }
+                //$this->Flash->error(__('The orcamento could not be saved. Please, try again.'));
             }
-            $this->Flash->error(__('The orcamento could not be saved. Please, try again.'));
+            $orcamentos = $this->Orcamentos->find()->where(["projeto_id"=>$projeto_id])->contain(['Projetos']);
         }
-        $projetos = $this->Orcamentos->Projetos->find('list', ['limit' => 200]);
-        $this->set(compact('orcamento', 'projetos'));
-    }
 
-    /**
-     * Edit method
-     *
-     * @param string|null $id Orcamento id.
-     * @return \Cake\Http\Response|null Redirects on successful edit, renders view otherwise.
-     * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
-     */
-    public function edit($id = null)
-    {
-        $orcamento = $this->Orcamentos->get($id, [
-            'contain' => []
-        ]);
-        if ($this->request->is(['patch', 'post', 'put'])) {
-            $orcamento = $this->Orcamentos->patchEntity($orcamento, $this->request->getData());
-            if ($this->Orcamentos->save($orcamento)) {
-                $this->Flash->success(__('The orcamento has been saved.'));
 
-                return $this->redirect(['action' => 'index']);
-            }
-            $this->Flash->error(__('The orcamento could not be saved. Please, try again.'));
-        }
-        $projetos = $this->Orcamentos->Projetos->find('list', ['limit' => 200]);
-        $this->set(compact('orcamento', 'projetos'));
+        $this->set(compact('orcamento','orcamentos','projeto_id'));
     }
 
     /**
@@ -96,16 +95,22 @@ class OrcamentosController extends AppController
      * @return \Cake\Http\Response|null Redirects to index.
      * @throws \Cake\Datasource\Exception\RecordNotFoundException When record not found.
      */
-    public function delete($id = null)
+    public function delete($id = null,$projeto_id=null)
     {
-        $this->request->allowMethod(['post', 'delete']);
-        $orcamento = $this->Orcamentos->get($id);
-        if ($this->Orcamentos->delete($orcamento)) {
-            $this->Flash->success(__('The orcamento has been deleted.'));
-        } else {
-            $this->Flash->error(__('The orcamento could not be deleted. Please, try again.'));
+        if(empty($id) || empty($projeto_id)){
+            throw new NotFoundException('Selecione o orçamento a ser excluído');
         }
 
-        return $this->redirect(['action' => 'index']);
+        $this->request->allowMethod(['post', 'delete']);
+        $orcamento = $this->Orcamentos->get($id);
+
+        if ($this->Orcamentos->delete($orcamento)) {
+            //$this->Flash->success(__('The renda has been deleted.'));
+        } else {
+            //$this->Flash->error(__('The renda could not be deleted. Please, try again.'));
+        }
+        $this->set('projeto_id',$projeto_id);
+
+        return $this->redirect(['action' => 'add',$projeto_id]);
     }
 }
